Run tests that use cloud_identity_group serially (#7322)
